+++ app.py
@@ -625,652 +625,13 @@ def bulk_action() -> Response:
 
     return redirect(url_for('index'))
 
-@app.route('/set_theme', methods=['POST'])
-def set_theme() -> Response:
-    """Store the selected theme in the session."""
-    theme = request.form.get('theme', '')
-    if theme in AVAILABLE_THEMES:
-        session['theme'] = theme
-        flash(f"Theme changed to '{theme.replace('theme-', '').replace('.css', '').capitalize()}'", "success")
-    else:
-        flash("Invalid theme selection.", "error")
-    return redirect(url_for('index'))
 
 
-@app.route('/set_background', methods=['POST'])
-def set_background() -> Response:
-    """Persist the selected background image in the session."""
-    bg = request.form.get('background', '')
-    if bg in AVAILABLE_BACKGROUNDS:
-        session['background'] = bg
-        return ('', 204)
-    return ('Invalid background', 400)
-
-
-@app.route('/set_panel_opacity', methods=['POST'])
-def set_panel_opacity() -> Response:
-    """Update and store the UI panel opacity setting."""
-    try:
-        opacity = float(request.form.get('opacity', '1'))
-    except ValueError:
-        return ('Invalid value', 400)
-    opacity = max(0.1, min(opacity, 1.0))
-    session['panel_opacity'] = opacity
-    return ('', 204)
-
-
-@app.route('/set_font_size', methods=['POST'])
-def set_font_size() -> Response:
-    """Update font size in the selected theme CSS file."""
-    try:
-        size = int(request.form.get('size', '14'))
-    except ValueError:
-        return ('Invalid value', 400)
-    size = max(10, min(size, 18))
-    theme = request.form.get('theme') or session.get('theme')
-    if not theme and AVAILABLE_THEMES:
-        theme = AVAILABLE_THEMES[0]
-    if not theme or theme not in AVAILABLE_THEMES:
-        return ('Invalid theme', 400)
-
-    path = os.path.join(THEMES_DIR, theme)
-    try:
-        lines = open(path).read().splitlines()
-    except OSError:
-        return ('Theme not found', 404)
-
-    new_lines = []
-    in_pagination = False
-    in_footer = False
-    for line in lines:
-        stripped = line.strip()
-        if stripped.startswith('.retrorecon-root .pagination'):
-            in_pagination = True
-        elif stripped.startswith('.retrorecon-root .footer'):
-            in_footer = True
-
-        if 'font-size:' in line:
-            if '13.333px' in line:
-                line = re.sub(r'font-size:\s*[^;]+;', f'font-size: {size}px;', line)
-            elif in_pagination or in_footer:
-                line = re.sub(r'font-size:\s*[^;]+;', f'font-size: {size}px;', line)
-
-        if in_pagination and '}' in line:
-            in_pagination = False
-        if in_footer and '}' in line:
-            in_footer = False
-        new_lines.append(line)
-
-    with open(path, 'w') as fh:
-        fh.write('\n'.join(new_lines) + '\n')
-    return ('', 204)
-
-
-@app.route('/set_items_per_page', methods=['POST'])
-def set_items_per_page() -> Response:
-    """Persist the preferred results-per-page value in the session."""
-    try:
-        count = int(request.form.get('count', ''))
-    except ValueError:
-        return ('Invalid value', 400)
-    if count not in ITEMS_PER_PAGE_OPTIONS:
-        return ('Invalid value', 400)
-    session['items_per_page'] = count
-    return redirect(url_for('index'))
-
-
-@app.route('/saved_tags', methods=['GET', 'POST'])
-def saved_tags_route() -> Response:
-    """Return or update the list of saved search tags."""
-
-    if request.method == 'GET':
-        return jsonify({'tags': load_saved_tags()})
-
-    tag = request.form.get('tag', '').strip()
-    if not tag:
-        return ('', 400)
-    if not tag.startswith('#'):
-        tag = '#' + tag
-    tags = load_saved_tags()
-    if tag not in tags:
-        tags.append(tag)
-        save_saved_tags(tags)
-    return ('', 204)
-
-
-@app.route('/delete_saved_tag', methods=['POST'])
-def delete_saved_tag() -> Response:
-    """Remove a tag from the saved list."""
-
-    tag = request.form.get('tag', '').strip()
-    if not tag:
-        return ('', 400)
-    if not tag.startswith('#'):
-        tag = '#' + tag
-    tags = load_saved_tags()
-    if tag in tags:
-        tags.remove(tag)
-        save_saved_tags(tags)
-    return ('', 204)
-
-
-@app.route('/notes/<int:url_id>', methods=['GET'])
-def notes_get(url_id: int) -> Response:
-    """Return all notes for the given URL as JSON."""
-
-    if not _db_loaded():
-        return jsonify([])
-    rows = get_notes(url_id)
-    return jsonify([
-        {
-            'id': r['id'],
-            'url_id': r['url_id'],
-            'content': r['content'],
-            'created_at': r['created_at'],
-            'updated_at': r['updated_at'],
-        }
-        for r in rows
-    ])
-
-
-@app.route('/notes', methods=['POST'])
-def notes_post() -> Response:
-    """Create or update a note."""
-
-    if not _db_loaded():
-        return ('', 400)
-    url_id = request.form.get('url_id', type=int)
-    content = request.form.get('content', '').strip()
-    if not url_id or not content:
-        return ('', 400)
-    note_id = request.form.get('note_id', type=int)
-    if note_id:
-        update_note(note_id, content)
-    else:
-        add_note(url_id, content)
-    return ('', 204)
-
-
-@app.route('/delete_note', methods=['POST'])
-def delete_note_route() -> Response:
-    """Delete a specific note or all notes for a URL."""
-
-    note_id = request.form.get('note_id', type=int)
-    url_id = request.form.get('url_id', type=int)
-    delete_all = request.form.get('all', '0') == '1'
-    if note_id:
-        delete_note_entry(note_id)
-    elif url_id and delete_all:
-        delete_all_notes(url_id)
-    else:
-        return ('', 400)
-    return ('', 204)
-
-
-@app.route('/export_notes', methods=['GET'])
-def export_notes() -> Response:
-    """Return all notes as JSON grouped by URL."""
-
-    if not _db_loaded():
-        return jsonify([])
-    data = export_notes_data()
-    return jsonify(data)
-
-
-@app.route('/text_tools', methods=['GET'])
-def text_tools_page() -> str:
-    """Return the Text Tools overlay HTML fragment."""
-
-    return render_template('text_tools.html')
-
-
-def _get_text_param() -> Optional[str]:
-    text = request.form.get('text', '')
-    if len(text.encode('utf-8')) > TEXT_TOOLS_LIMIT:
-        return None
-    return text
-
-
-@app.route('/tools/base64_decode', methods=['POST'])
-def base64_decode_route() -> Response:
-    """Decode Base64 text."""
-
-    text = _get_text_param()
-    if text is None:
-        return ('Request too large', 400)
-    import base64
-    try:
-        decoded = base64.b64decode(text, validate=True)
-        decoded.decode('utf-8')  # ensure text only
-    except Exception:
-        return ('Invalid Base64 data', 400)
-    return Response(decoded, mimetype='text/plain')
-
-
-@app.route('/tools/base64_encode', methods=['POST'])
-def base64_encode_route() -> Response:
-    """Encode text as Base64."""
-
-    text = _get_text_param()
-    if text is None:
-        return ('Request too large', 400)
-    import base64
-    encoded = base64.b64encode(text.encode('utf-8')).decode('ascii')
-    return Response(encoded, mimetype='text/plain')
-
-
-@app.route('/tools/url_decode', methods=['POST'])
-def url_decode_route() -> Response:
-    """URL-decode percent-encoded text."""
-
-    text = _get_text_param()
-    if text is None:
-        return ('Request too large', 400)
-    try:
-        decoded = urllib.parse.unquote_plus(text)
-    except Exception:
-        return ('Invalid URL encoding', 400)
-    return Response(decoded, mimetype='text/plain')
-
-
-@app.route('/tools/url_encode', methods=['POST'])
-def url_encode_route() -> Response:
-    """Percent-encode text for URLs."""
-
-    text = _get_text_param()
-    if text is None:
-        return ('Request too large', 400)
-    encoded = urllib.parse.quote_plus(text)
-    return Response(encoded, mimetype='text/plain')
-
-
-@app.route('/jwt_tools', methods=['GET'])
-def jwt_tools_page() -> str:
-    """Return the JWT Tools overlay HTML fragment."""
-
-    return render_template('jwt_tools.html')
-
-
-@app.route('/tools/jwt', methods=['GET'])
-def jwt_tools_full_page() -> str:
-    """Render index page with JWT Tools open for bookmarkable view."""
-
-    return index()
-
-
-@app.route('/tools/jwt_decode', methods=['POST'])
-def jwt_decode_route() -> Response:
-    """Decode a JWT without verifying the signature."""
-
-    token = request.form.get('token', '')
-    if len(token.encode('utf-8')) > TEXT_TOOLS_LIMIT:
-        return ('Request too large', 400)
-    if not _db_loaded():
-        return jsonify({'error': 'no_db'}), 400
-    try:
-        header = jwt.get_unverified_header(token)
-        payload = jwt.decode(token, options={'verify_signature': False})
-    except Exception as e:
-        return (f'Invalid JWT: {e}', 400)
-
-    result: Dict[str, Any] = {"header": header, "payload": payload}
-
-    # Timestamp conversion
-    now = datetime.datetime.now(datetime.UTC).timestamp()
-    if isinstance(payload, dict):
-        for field in ["iat", "nbf", "exp"]:
-            val = payload.get(field)
-            if isinstance(val, (int, float)):
-                dt = datetime.datetime.fromtimestamp(val, datetime.UTC)
-                result[f"{field}_readable"] = dt.strftime("%Y-%m-%d %H:%M:%S")
-    exp_val = payload.get("exp")
-    if isinstance(exp_val, (int, float)):
-        result["expired"] = now > exp_val
-        result["exp_readable"] = datetime.datetime.fromtimestamp(exp_val, datetime.UTC).strftime("%Y-%m-%d %H:%M:%S")
-    else:
-        result["expired"] = False
-
-    alg = header.get("alg", "").lower()
-    weak_algs = {"none", "hs256", "hs384", "hs512"}
-    result["alg_warning"] = alg in weak_algs
-
-    key_warning = False
-    known_keys = ["secret", "secret123", "changeme", "password"]
-    for k in known_keys:
-        try:
-            jwt.decode(
-                token,
-                k,
-                algorithms=[header.get("alg", "")],
-                options={"verify_signature": True, "verify_exp": False},
-            )
-            key_warning = True
-            break
-        except jwt.InvalidSignatureError:
-            continue
-        except Exception:
-            continue
-    result["key_warning"] = key_warning
-
-    notes = []
-    if result.get("alg_warning"):
-        notes.append("Weak algorithm")
-    if result.get("key_warning"):
-        notes.append("Weak key")
-    if result.get("exp_readable"):
-        note = f"exp {result['exp_readable']}"
-        if result.get("expired"):
-            note += " expired"
-        notes.append(note)
-    try:
-        log_jwt_entry(token, header, payload, "; ".join(notes))
-    except Exception:
-        pass
-
-    return jsonify(result)
-
-
-@app.route('/tools/jwt_encode', methods=['POST'])
-def jwt_encode_route() -> Response:
-    """Encode JSON payload into a JWT."""
-
-    raw = request.form.get('payload', '')
-    if len(str(raw).encode('utf-8')) > TEXT_TOOLS_LIMIT:
-        return ('Request too large', 400)
-    secret = request.form.get('secret') or None
-
-    def _parse(text: str):
-        try:
-            return json.loads(text)
-        except Exception:
-            try:
-                import ast
-                return ast.literal_eval(text)
-            except Exception:
-                return None
-
-    payload = _parse(raw)
-    if payload is None:
-        for sep in ['}\n{', '}\r\n{']:
-            if sep in raw:
-                candidate = '{' + raw.split(sep, 1)[1]
-                payload = _parse(candidate)
-                if payload is not None:
-                    break
-    if payload is None:
-        payload = {}
-    if secret:
-        token = jwt.encode(payload, secret, algorithm='HS256')
-    else:
-        token = jwt.encode(payload, '', algorithm='none')
-    if isinstance(token, bytes):
-        token = token.decode('ascii')
-    return Response(token, mimetype='text/plain')
-
-
-@app.route('/jwt_cookies', methods=['GET'])
-def jwt_cookies_route() -> Response:
-    """Return logged JWT decode entries as JSON."""
-
-    if not _db_loaded():
-        return jsonify([])
-    rows = export_jwt_cookie_data()[:50]
-    return jsonify(rows)
-
-
-@app.route('/delete_jwt_cookies', methods=['POST'])
-def delete_jwt_cookies_route() -> Response:
-    """Delete JWT cookie entries by ID."""
-
-    if not _db_loaded():
-        return ('', 400)
-    ids = [int(i) for i in request.form.getlist('ids') if i.isdigit()]
-    if not ids:
-        return ('', 400)
-    delete_jwt_cookies(ids)
-    return ('', 204)
-
-
-@app.route('/update_jwt_cookie', methods=['POST'])
-def update_jwt_cookie_route() -> Response:
-    """Update notes for a JWT cookie entry."""
-
-    if not _db_loaded():
-        return ('', 400)
-    jid = request.form.get('id', type=int)
-    notes = request.form.get('notes', '').strip()
-    if not jid:
-        return ('', 400)
-    update_jwt_cookie(jid, notes)
-    return ('', 204)
-
-
-@app.route('/export_jwt_cookies', methods=['GET'])
-def export_jwt_cookies_route() -> Response:
-    """Return JWT cookie entries as JSON."""
-
-    if not _db_loaded():
-        return jsonify([])
-    ids = [int(i) for i in request.args.getlist('id') if i.isdigit()]
-    rows = export_jwt_cookie_data(ids if ids else None)
-    return jsonify(rows)
-
-
-@app.route('/screenshotter', methods=['GET'])
-def screenshotter_page() -> str:
-    """Return the ScreenShotter overlay HTML fragment."""
-
-    return render_template('screenshotter.html')
-
-
-@app.route('/tools/screenshotter', methods=['GET'])
-def screenshotter_full_page() -> str:
-    """Render index page with ScreenShotter open for bookmarkable view."""
-
-    return index()
-
-
-@app.route('/tools/screenshot', methods=['POST'])
-def screenshot_route() -> Response:
-    """Capture a screenshot of a webpage and log it."""
-
-    if not _db_loaded():
-        return jsonify({'error': 'no_db'}), 400
-    url = request.form.get('url', '').strip()
-    if not url:
-        return ('Missing URL', 400)
-    agent = request.form.get('agent', '')
-    spoof = request.form.get('spoof', '0') == '1'
-    try:
-        img_bytes = take_screenshot(url, agent, spoof)
-    except Exception as e:
-        return (f'Error taking screenshot: {e}', 500)
-    ts = int(datetime.datetime.now(datetime.UTC).timestamp() * 1000)
-    fname = f'shot_{ts}.png'
-    thumb = f'shot_{ts}_th.png'
-    os.makedirs(SCREENSHOT_DIR, exist_ok=True)
-    full_path = os.path.join(SCREENSHOT_DIR, fname)
-    with open(full_path, 'wb') as f:
-        f.write(img_bytes)
-    thumb_path = os.path.join(SCREENSHOT_DIR, thumb)
-    try:
-        from PIL import Image
-        img = Image.open(io.BytesIO(img_bytes))
-        img.thumbnail((160, 160))
-        img.save(thumb_path, format='PNG')
-    except Exception as e:  # pragma: no cover - pillow might not be installed
-        logger.debug('thumbnail generation failed: %s', e)
-        with open(thumb_path, 'wb') as f:
-            f.write(img_bytes)
-    sid = save_screenshot_record(url, fname, thumb, 'GET')
-    return jsonify({'id': sid})
-
-
-@app.route('/screenshots', methods=['GET'])
-def screenshots_route() -> Response:
-    """Return screenshot metadata as JSON."""
-
-    if not _db_loaded():
-        return jsonify([])
-    rows = list_screenshot_data()
-    for r in rows:
-        r['file'] = url_for('static', filename='screenshots/' + r['screenshot_path'])
-        r['preview'] = url_for('static', filename='screenshots/' + r['thumbnail_path'])
-    return jsonify(rows)
-
-
-@app.route('/delete_screenshots', methods=['POST'])
-def delete_screenshots_route() -> Response:
-    """Delete screenshot entries by ID."""
-
-    if not _db_loaded():
-        return ('', 400)
-    ids = [int(i) for i in request.form.getlist('ids') if i.isdigit()]
-    if not ids:
-        return ('', 400)
-    delete_screenshots(ids)
-    return ('', 204)
-
-@app.route('/tools/webpack-zip', methods=['POST'])
-def webpack_zip() -> Response:
-    """Package sources from a Webpack ``.map`` file into a ZIP archive."""
-    map_url = request.form.get('map_url', '').strip()
-    if not map_url:
-        flash("No .js.map URL provided.", "error")
-        return redirect(url_for('index'))
-
-    try:
-        resp = requests.get(map_url, timeout=15)
-        resp.raise_for_status()
-        raw_text = resp.text
-
-        map_data = json.loads(raw_text)
-
-        sources = map_data.get('sources', [])
-        sources_content = map_data.get('sourcesContent', [])
-
-        if len(sources_content) != len(sources):
-            flash(
-                "Warning: sourcesContent length does not match sources length. "
-                "Some files may be skipped.",
-                "warning"
-            )
-
-        zip_buffer = io.BytesIO()
-        with zipfile.ZipFile(zip_buffer, 'w', zipfile.ZIP_DEFLATED) as zipf:
-            for idx, src_path in enumerate(sources):
-                filename_in_zip = src_path.replace('\\', '/').lstrip('/')
-                if idx < len(sources_content) and sources_content[idx] is not None:
-                    file_bytes = sources_content[idx].encode('utf-8')
-                else:
-                    continue
-                zipf.writestr(filename_in_zip, file_bytes)
-
-        zip_buffer.seek(0)
-        original_map_name = map_url.rstrip('/').split('/')[-1]
-        zip_filename = original_map_name.replace('.map', '.zip')
-
-        return send_file(
-            zip_buffer,
-            mimetype='application/zip',
-            as_attachment=True,
-            download_name=zip_filename
-        )
-
-    except requests.exceptions.RequestException as e:
-        flash(f"Error fetching .js.map: {e}", "error")
-        return redirect(url_for('index'))
-
-    except (json.JSONDecodeError, KeyError) as e:
-        flash(f"Error parsing .js.map JSON: {e}", "error")
-        return redirect(url_for('index'))
-
-    except Exception as e:
-        flash(f"Unexpected server error: {e}", "error")
-        return redirect(url_for('index'))
-
-@app.route('/new_db', methods=['POST'])
-def new_db() -> Response:
-    """Create a fresh empty database."""
-    name = request.form.get('db_name', '').strip()
-    safe = _sanitize_db_name(name)
-    if not safe:
-        flash('Invalid database name.', 'error')
-        return redirect(url_for('index'))
-    close_connection(None)
-    try:
-        db_name = create_new_db(safe)
-        session['db_display_name'] = db_name
-        flash('New database created.', 'success')
-    except ValueError as e:
-        flash(str(e), 'error')
-    return redirect(url_for('index'))
-
-@app.route('/load_db', methods=['POST'])
-def load_db_route() -> Response:
-    """Upload and load a database file provided by the user."""
-    file = request.files.get('db_file')
-    if not file:
-        flash("No database file uploaded.", "error")
-        return redirect(url_for('index'))
-    filename = _sanitize_db_name(file.filename or '')
-    if not filename:
-        flash('Invalid database file.', 'error')
-        return redirect(url_for('index'))
-    db_path = os.path.join(app.root_path, filename)
-    close_connection(None)
-    try:
-        file.save(db_path)
-        app.config['DATABASE'] = db_path
-        ensure_schema()
-        session['db_display_name'] = filename
-        flash("Database loaded.", "success")
-    except Exception as e:
-        flash(f"Error loading database: {e}", "error")
-    return redirect(url_for('index'))
-
-@app.route('/save_db', methods=['GET'])
-def save_db() -> Response:
-    """Return the database file for download."""
-
-    if not _db_loaded():
-        flash('No database loaded.', 'error')
-        return redirect(url_for('index'))
-    name = request.args.get("name", "").strip()
-    if name:
-        safe_name = _sanitize_export_name(name)
-    else:
-        safe_name = os.path.basename(app.config["DATABASE"])
-    return send_file(
-        app.config['DATABASE'],
-        as_attachment=True,
-        download_name=safe_name
-    )
-
-
-@app.route('/rename_db', methods=['POST'])
-def rename_db() -> Response:
-    """Rename the current database file."""
-    new_name = request.form.get('new_name', '').strip()
-    safe = _sanitize_db_name(new_name or '')
-    if not safe:
-        flash('Invalid database name.', 'error')
-        return redirect(url_for('index'))
-    if not _db_loaded():
-        flash('No database loaded.', 'error')
-        return redirect(url_for('index'))
-    close_connection(None)
-    new_path = os.path.join(app.root_path, safe)
-    try:
-        os.rename(app.config['DATABASE'], new_path)
-    except OSError as e:
-        flash(f'Error renaming database: {e}', 'error')
-        return redirect(url_for('index'))
-    app.config['DATABASE'] = new_path
-    ensure_schema()
-    session['db_display_name'] = safe
-    flash('Database renamed.', 'success')
-    return redirect(url_for('index'))
+from retrorecon.routes import notes_bp, tools_bp, db_bp, settings_bp
+app.register_blueprint(notes_bp)
+app.register_blueprint(tools_bp)
+app.register_blueprint(db_bp)
+app.register_blueprint(settings_bp)
 
 if __name__ == '__main__':
     if env_db and app.config.get('DATABASE'):
